rabbitmq: block client port on startup
This script blocks the connection to the rabbitmq cluster in case the number of nodes decay bellow the half
of the total. In this case the remain rabbit nodes won't accept new connections if the quorum is reached.

It take advantage of the pacemaker notifications that notify when a rabbitmq has failed or has restored.
All nodes rabbitmq ports will be blocked if the total number of alive nodes are below to the half of
nodes of the cluster, or unblock if its over this value.

The script is divided in two parts: On one hand, the alert handler that manages the alerts discarding non
interesting alerts, launches to the blocker script in this node and other cluster nodes (via SSH).
And in the other hand, the blocker script than checks the condition of running nodes and blocks or unblocks
the rabbitmq client port.

co-authored-by: Ivan Lausuch ilausuch@suse.com